ai-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
groq==0.4.1
redis==5.0.1
python-dotenv==1.0.0
